Replace `CollectionsMarshal.AsSpan` with `IList<T>` indexing in `ToFastArray`
