+++ src/core/StackExchange.Redis.Extensions.Core/Implementations/RedisDatabase.HyperLogLog.cs
@@ -3,7 +3,7 @@
 using System;
 using System.Threading.Tasks;
 
-using StackExchange.Redis.Extensions.Core.Helpers;
+using StackExchange.Redis.Extensions.Core.Extensions;
 
 namespace StackExchange.Redis.Extensions.Core.Implementations;
