+++ src/core/StackExchange.Redis.Extensions.Core/Implementations/RedisDatabase.List.cs
@@ -4,7 +4,7 @@ using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
-using StackExchange.Redis.Extensions.Core.Helpers;
+using StackExchange.Redis.Extensions.Core.Extensions;
 
 namespace StackExchange.Redis.Extensions.Core.Implementations;
