+++ src/core/StackExchange.Redis.Extensions.Core/Implementations/RedisDatabase.PubSub.cs
@@ -4,7 +4,7 @@ using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
-using StackExchange.Redis.Extensions.Core.Helpers;
+using StackExchange.Redis.Extensions.Core.Extensions;
 using StackExchange.Redis.Extensions.Core.Logging;
 
 namespace StackExchange.Redis.Extensions.Core.Implementations;
