+++ tests/StackExchange.Redis.Extensions.Core.Tests/CacheClientTestBase.Hash.cs
@@ -4,7 +4,7 @@ using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
-using StackExchange.Redis.Extensions.Core.Helpers;
+using StackExchange.Redis.Extensions.Core.Extensions;
 using StackExchange.Redis.Extensions.Core.Tests.Helpers;
 
 using Xunit;
